Warning fix

+++ MoPubSDK/Internal/Common/AdAlerts/MPAdAlertGestureRecognizer.m
@@ -238,7 +238,7 @@ NSInteger const kMPAdAlertGestureMaxAllowedYAxisMovement = 50;
 {
     CGPoint nowPoint = [touches.anyObject locationInView:self.view];
     
-    return fabsf(nowPoint.y - self.startingPoint.y) <= kMPAdAlertGestureMaxAllowedYAxisMovement;
+    return fabs(nowPoint.y - self.startingPoint.y) <= kMPAdAlertGestureMaxAllowedYAxisMovement;
 }
 
 - (BOOL)touchIsWithinBoundsForTouches:(NSSet *)touches
